Move the client configuration log from the gapic trait to the options trait
